#107 Conversion of the sorting and filter function.
In the respective view models, only the list and a sorting or filter option (string) are passed for sorting or filtering. Sorting or filtering is done in the Utils class ArrayListUtil (formerly CollectionsSorter). The desired sorting or filtering function is executed there using switch. The purpose of this is to make the ViewModels tidier and clearer. Currently creation times, like count, response count and response status are covered. The rest is following.