#206 filter functions
Filter functions now work. If filtering is switched off, the full list is displayed again. It can be filtered and sorted at the same time. However, only one filter function can be active at the same time because, for example, the simultaneous filtering for both self-created threads and threads created by the course creator makes no sense because the result would be an empty list.